refactor: strict typing for openapi3 and swagger
