+++ src/openapi3/handleJson.test.ts
@@ -1,6 +1,8 @@
+import { OpenAPIV3 } from "openapi-types";
+import { Field } from "../";
 import handleJson from "./handleJson";
 
-const swaggerApiDefinition = {
+const openApi3Definition: OpenAPIV3.Document = {
   openapi: "3.0.2",
   info: {
     title: "",
@@ -1358,7 +1360,7 @@ const parsed = [
 
 describe(`Parse OpenApi v3 Documentation from Json`, () => {
   const toBeParsed = handleJson(
-    swaggerApiDefinition,
+    openApi3Definition,
     "https://demo.api-platform.com"
   );
 
@@ -1371,6 +1373,6 @@ describe(`Parse OpenApi v3 Documentation from Json`, () => {
     expect(toBeParsed[1].url).toBe(parsed[1].url);
     expect(toBeParsed[1].id).toBe(parsed[1].id);
 
-    expect(toBeParsed[1].fields[0]).toEqual(parsed[1].fields[0]);
+    expect((toBeParsed[1].fields as Field[])[0]).toEqual(parsed[1].fields[0]);
   });
 });
