+++ src/swagger/index.ts
@@ -1,3 +1,1 @@
-export {
-  default as parseSwaggerDocumentation
-} from "./parseSwaggerDocumentation";
+export * from "./parseSwaggerDocumentation";
