+++ src/swagger/parseSwaggerDocumentation.ts
@@ -1,7 +1,16 @@
-import Api from "../Api";
+import { OpenAPIV2 } from "openapi-types";
+import { Api } from "../Api";
 import handleJson, { removeTrailingSlash } from "./handleJson";
 
-export default function parseSwaggerDocumentation(entrypointUrl: string) {
+export interface ParsedSwaggerDocumentation {
+  api: Api;
+  response: OpenAPIV2.Document;
+  status: string;
+}
+
+export function parseSwaggerDocumentation(
+  entrypointUrl: string
+): Promise<ParsedSwaggerDocumentation> {
   entrypointUrl = removeTrailingSlash(entrypointUrl);
   return fetch(entrypointUrl)
     .then(res => res.json())
